+++ packages/amplify-category-api/src/graphql-transformer/transform-graphql-schema.ts
@@ -1,6 +1,6 @@
-import { DeploymentResources as DeploymentResourcesV2 } from '@aws-amplify/graphql-transformer-core';
-import { $TSAny, $TSContext, ApiCategoryFacade } from 'amplify-cli-core';
+import { DeploymentResources as DeploymentResourcesV2 } from '@aws-amplify/graphql-transformer-interfaces';
 import { DeploymentResources as DeploymentResourcesV1 } from 'graphql-transformer-core';
+import { $TSAny, $TSContext, ApiCategoryFacade } from 'amplify-cli-core';
 import { AmplifyGraphQLTransformerErrorConverter } from '../errors/amplify-error-converter';
 import { transformGraphQLSchemaV1 } from './transform-graphql-schema-v1';
 import { transformGraphQLSchemaV2 } from './transform-graphql-schema-v2';