+++ packages/amplify-graphql-function-transformer/src/__tests__/amplify-graphql-function-transformer.test.ts
@@ -1,5 +1,5 @@
 'use strict';
-import { anything, countResources, expect as cdkExpect, haveResource } from '@aws-cdk/assert';
+import { Match, Template } from 'aws-cdk-lib/assertions';
 import { GraphQLTransform } from '@aws-amplify/graphql-transformer-core';
 import { parse } from 'graphql';
 import { FunctionTransformer } from '..';
@@ -21,13 +21,13 @@ test('for @function with only name, it generates the expected resources', () => 
   parse(out.schema);
   const stack = out.stacks.FunctionDirectiveStack;
   expect(stack).toBeDefined();
-  cdkExpect(stack).to(countResources('AWS::IAM::Role', 1));
-  cdkExpect(stack).to(countResources('AWS::IAM::Policy', 1));
-  cdkExpect(stack).to(countResources('AWS::AppSync::DataSource', 1));
-  cdkExpect(stack).to(countResources('AWS::AppSync::FunctionConfiguration', 1));
-  cdkExpect(stack).to(countResources('AWS::AppSync::Resolver', 1));
-  cdkExpect(stack).to(
-    haveResource('AWS::IAM::Role', {
+  Template.fromJSON(stack).resourceCountIs('AWS::IAM::Role', 1);
+  Template.fromJSON(stack).resourceCountIs('AWS::IAM::Policy', 1);
+  Template.fromJSON(stack).resourceCountIs('AWS::AppSync::DataSource', 1);
+  Template.fromJSON(stack).resourceCountIs('AWS::AppSync::FunctionConfiguration', 1);
+  Template.fromJSON(stack).resourceCountIs('AWS::AppSync::Resolver', 1);
+  Template.fromJSON(stack)
+    .hasResourceProperties('AWS::IAM::Role', {
       AssumeRolePolicyDocument: {
         Statement: [
           {
@@ -40,10 +40,9 @@ test('for @function with only name, it generates the expected resources', () => 
         ],
         Version: '2012-10-17',
       },
-    }),
-  );
-  cdkExpect(stack).to(
-    haveResource('AWS::IAM::Policy', {
+    });
+  Template.fromJSON(stack)
+    .hasResourceProperties('AWS::IAM::Policy', {
       PolicyDocument: {
         Statement: [
           {
@@ -54,7 +53,7 @@ test('for @function with only name, it generates the expected resources', () => 
                 'Fn::If': [
                   'HasEnvironmentParameter',
                   {
-                    'Fn::Sub': ['arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:echofunction-${env}', { env: { Ref: anything() } }],
+                    'Fn::Sub': ['arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:echofunction-${env}', { env: { Ref: Match.anyValue() } }],
                   },
                   { 'Fn::Sub': 'arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:echofunction' },
                 ],
@@ -67,7 +66,7 @@ test('for @function with only name, it generates the expected resources', () => 
                       'Fn::If': [
                         'HasEnvironmentParameter',
                         {
-                          'Fn::Sub': ['arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:echofunction-${env}', { env: { Ref: anything() } }],
+                          'Fn::Sub': ['arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:echofunction-${env}', { env: { Ref: Match.anyValue() } }],
                         },
                         { 'Fn::Sub': 'arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:echofunction' },
                       ],
@@ -76,18 +75,17 @@ test('for @function with only name, it generates the expected resources', () => 
                   ],
                 ],
               },
-          ],
+            ],
           },
         ],
         Version: '2012-10-17',
       },
-      PolicyName: anything(),
-      Roles: [{ Ref: anything() }],
-    }),
-  );
-  cdkExpect(stack).to(
-    haveResource('AWS::AppSync::DataSource', {
-      ApiId: { Ref: anything() },
+      PolicyName: Match.anyValue(),
+      Roles: [{ Ref: Match.anyValue() }],
+    });
+  Template.fromJSON(stack)
+    .hasResourceProperties('AWS::AppSync::DataSource', {
+      ApiId: { Ref: Match.anyValue() },
       Name: 'EchofunctionLambdaDataSource',
       Type: 'AWS_LAMBDA',
       LambdaConfig: {
@@ -95,7 +93,7 @@ test('for @function with only name, it generates the expected resources', () => 
           'Fn::If': [
             'HasEnvironmentParameter',
             {
-              'Fn::Sub': ['arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:echofunction-${env}', { env: { Ref: anything() } }],
+              'Fn::Sub': ['arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:echofunction-${env}', { env: { Ref: Match.anyValue() } }],
             },
             { 'Fn::Sub': 'arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:echofunction' },
           ],
@@ -104,37 +102,34 @@ test('for @function with only name, it generates the expected resources', () => 
       ServiceRoleArn: {
         'Fn::GetAtt': ['EchofunctionLambdaDataSourceServiceRole3BE2FA57', 'Arn'],
       },
-    }),
-  );
-  cdkExpect(stack).to(
-    haveResource('AWS::AppSync::FunctionConfiguration', {
-      ApiId: { Ref: anything() },
-      DataSourceName: { 'Fn::GetAtt': [anything(), 'Name'] },
+    });
+  Template.fromJSON(stack)
+    .hasResourceProperties('AWS::AppSync::FunctionConfiguration', {
+      ApiId: { Ref: Match.anyValue() },
+      DataSourceName: { 'Fn::GetAtt': [Match.anyValue(), 'Name'] },
       FunctionVersion: '2018-05-29',
       Name: 'InvokeEchofunctionLambdaDataSource',
       RequestMappingTemplateS3Location: {
-        'Fn::Join': ['', ['s3://', { Ref: anything() }, '/', { Ref: anything() }, '/resolvers/InvokeEchofunctionLambdaDataSource.req.vtl']],
+        'Fn::Join': ['', ['s3://', { Ref: Match.anyValue() }, '/', { Ref: Match.anyValue() }, '/resolvers/InvokeEchofunctionLambdaDataSource.req.vtl']],
       },
       ResponseMappingTemplateS3Location: {
-        'Fn::Join': ['', ['s3://', { Ref: anything() }, '/', { Ref: anything() }, '/resolvers/InvokeEchofunctionLambdaDataSource.res.vtl']],
+        'Fn::Join': ['', ['s3://', { Ref: Match.anyValue() }, '/', { Ref: Match.anyValue() }, '/resolvers/InvokeEchofunctionLambdaDataSource.res.vtl']],
       },
-    }),
-  );
-  cdkExpect(stack).to(
-    haveResource('AWS::AppSync::Resolver', {
-      ApiId: { Ref: anything() },
+    });
+  Template.fromJSON(stack)
+    .hasResourceProperties('AWS::AppSync::Resolver', {
+      ApiId: { Ref: Match.anyValue() },
       FieldName: 'echo',
       TypeName: 'Query',
       Kind: 'PIPELINE',
       PipelineConfig: {
-        Functions: [{ 'Fn::GetAtt': [anything(), 'FunctionId'] }],
+        Functions: [{ 'Fn::GetAtt': [Match.anyValue(), 'FunctionId'] }],
       },
-      RequestMappingTemplate: anything(),
+      RequestMappingTemplate: Match.anyValue(),
       ResponseMappingTemplateS3Location: {
-        'Fn::Join': ['', ['s3://', { Ref: anything() }, '/', { Ref: anything() }, '/resolvers/Query.echo.res.vtl']],
+        'Fn::Join': ['', ['s3://', { Ref: Match.anyValue() }, '/', { Ref: Match.anyValue() }, '/resolvers/Query.echo.res.vtl']],
       },
-    }),
-  );
+    });
   expect(out.resolvers).toMatchSnapshot();
 });
 
@@ -155,120 +150,110 @@ test('for @function with account ID, it generates the expected resources', () =>
   parse(out.schema);
   const stack = out.stacks.FunctionDirectiveStack;
   expect(stack).toBeDefined();
-  cdkExpect(stack).to(countResources('AWS::IAM::Role', 1));
-  cdkExpect(stack).to(countResources('AWS::IAM::Policy', 1));
-  cdkExpect(stack).to(countResources('AWS::AppSync::DataSource', 1));
-  cdkExpect(stack).to(countResources('AWS::AppSync::FunctionConfiguration', 1));
-  cdkExpect(stack).to(countResources('AWS::AppSync::Resolver', 1));
-  cdkExpect(stack).to(
-    haveResource('AWS::IAM::Role', {
-      AssumeRolePolicyDocument: {
-        Statement: [
-          {
-            Action: 'sts:AssumeRole',
-            Effect: 'Allow',
-            Principal: {
-              Service: 'appsync.amazonaws.com',
-            },
+  Template.fromJSON(stack).resourceCountIs('AWS::IAM::Role', 1);
+  Template.fromJSON(stack).resourceCountIs('AWS::IAM::Policy', 1);
+  Template.fromJSON(stack).resourceCountIs('AWS::AppSync::DataSource', 1);
+  Template.fromJSON(stack).resourceCountIs('AWS::AppSync::FunctionConfiguration', 1);
+  Template.fromJSON(stack).resourceCountIs('AWS::AppSync::Resolver', 1);
+  Template.fromJSON(stack).hasResourceProperties('AWS::IAM::Role', {
+    AssumeRolePolicyDocument: {
+      Statement: [
+        {
+          Action: 'sts:AssumeRole',
+          Effect: 'Allow',
+          Principal: {
+            Service: 'appsync.amazonaws.com',
           },
-        ],
-        Version: '2012-10-17',
-      },
-    }),
-  );
-  cdkExpect(stack).to(
-    haveResource('AWS::IAM::Policy', {
-      PolicyDocument: {
-        Statement: [
-          {
-            Action: 'lambda:InvokeFunction',
-            Effect: 'Allow',
-            Resource: [
-              {
-                'Fn::If': [
-                  'HasEnvironmentParameter',
-                  {
-                    'Fn::Sub': ['arn:aws:lambda:${AWS::Region}:123123456456:function:echofunction', {}],
-                  },
-                  { 'Fn::Sub': 'arn:aws:lambda:${AWS::Region}:123123456456:function:echofunction' },
-                ],
-              },
-              {
-                'Fn::Join': [
-                  '',
-                  [
-                    {
-                      'Fn::If': [
-                        'HasEnvironmentParameter',
-                        {
-                          'Fn::Sub': ['arn:aws:lambda:${AWS::Region}:123123456456:function:echofunction', {}],
-                        },
-                        { 'Fn::Sub': 'arn:aws:lambda:${AWS::Region}:123123456456:function:echofunction' },
-                      ],
-                    },
-                    ':*',
-                  ],
-                ],
-              },
-            ],
-          },
-        ],
-        Version: '2012-10-17',
-      },
-      PolicyName: anything(),
-      Roles: [{ Ref: anything() }],
-    }),
-  );
-  cdkExpect(stack).to(
-    haveResource('AWS::AppSync::DataSource', {
-      ApiId: { Ref: anything() },
-      Name: 'Echofunction123123456456LambdaDataSource',
-      Type: 'AWS_LAMBDA',
-      LambdaConfig: {
-        LambdaFunctionArn: {
-          'Fn::If': [
-            'HasEnvironmentParameter',
+        },
+      ],
+      Version: '2012-10-17',
+    },
+  });
+  Template.fromJSON(stack).hasResourceProperties('AWS::IAM::Policy', {
+    PolicyDocument: {
+      Statement: [
+        {
+          Action: 'lambda:InvokeFunction',
+          Effect: 'Allow',
+          Resource: [
             {
-              'Fn::Sub': ['arn:aws:lambda:${AWS::Region}:123123456456:function:echofunction', {}],
+              'Fn::If': [
+                'HasEnvironmentParameter',
+                {
+                  'Fn::Sub': ['arn:aws:lambda:${AWS::Region}:123123456456:function:echofunction', {}],
+                },
+                { 'Fn::Sub': 'arn:aws:lambda:${AWS::Region}:123123456456:function:echofunction' },
+              ],
             },
-            { 'Fn::Sub': 'arn:aws:lambda:${AWS::Region}:123123456456:function:echofunction' },
+            {
+              'Fn::Join': [
+                '',
+                [
+                  {
+                    'Fn::If': [
+                      'HasEnvironmentParameter',
+                      {
+                        'Fn::Sub': ['arn:aws:lambda:${AWS::Region}:123123456456:function:echofunction', {}],
+                      },
+                      { 'Fn::Sub': 'arn:aws:lambda:${AWS::Region}:123123456456:function:echofunction' },
+                    ],
+                  },
+                  ':*',
+                ],
+              ],
+            },
           ],
         },
+      ],
+      Version: '2012-10-17',
+    },
+    PolicyName: Match.anyValue(),
+    Roles: [{ Ref: Match.anyValue() }],
+  });
+  Template.fromJSON(stack).hasResourceProperties('AWS::AppSync::DataSource', {
+    ApiId: { Ref: Match.anyValue() },
+    Name: 'Echofunction123123456456LambdaDataSource',
+    Type: 'AWS_LAMBDA',
+    LambdaConfig: {
+      LambdaFunctionArn: {
+        'Fn::If': [
+          'HasEnvironmentParameter',
+          {
+            'Fn::Sub': ['arn:aws:lambda:${AWS::Region}:123123456456:function:echofunction', {}],
+          },
+          { 'Fn::Sub': 'arn:aws:lambda:${AWS::Region}:123123456456:function:echofunction' },
+        ],
       },
-      ServiceRoleArn: {
-        'Fn::GetAtt': ['Echofunction123123456456LambdaDataSourceServiceRole0B60B47E', 'Arn'],
-      },
-    }),
-  );
-  cdkExpect(stack).to(
-    haveResource('AWS::AppSync::FunctionConfiguration', {
-      ApiId: { Ref: anything() },
-      DataSourceName: { 'Fn::GetAtt': [anything(), 'Name'] },
-      FunctionVersion: '2018-05-29',
-      Name: 'InvokeEchofunction123123456456LambdaDataSource',
-      RequestMappingTemplateS3Location: {
-        'Fn::Join': ['', ['s3://', { Ref: anything() }, '/', { Ref: anything() }, '/resolvers/InvokeEchofunction123123456456LambdaDataSource.req.vtl']],
-      },
-      ResponseMappingTemplateS3Location: {
-        'Fn::Join': ['', ['s3://', { Ref: anything() }, '/', { Ref: anything() }, '/resolvers/InvokeEchofunction123123456456LambdaDataSource.res.vtl']],
-      },
-    }),
-  );
-  cdkExpect(stack).to(
-    haveResource('AWS::AppSync::Resolver', {
-      ApiId: { Ref: anything() },
-      FieldName: 'echo',
-      TypeName: 'Query',
-      Kind: 'PIPELINE',
-      PipelineConfig: {
-        Functions: [{ 'Fn::GetAtt': [anything(), 'FunctionId'] }],
-      },
-      RequestMappingTemplate: anything(),
-      ResponseMappingTemplateS3Location: {
-        'Fn::Join': ['', ['s3://', { Ref: anything() }, '/', { Ref: anything() }, '/resolvers/Query.echo.res.vtl']],
-      },
-    }),
-  );
+    },
+    ServiceRoleArn: {
+      'Fn::GetAtt': ['Echofunction123123456456LambdaDataSourceServiceRole0B60B47E', 'Arn'],
+    },
+  });
+  Template.fromJSON(stack).hasResourceProperties('AWS::AppSync::FunctionConfiguration', {
+    ApiId: { Ref: Match.anyValue() },
+    DataSourceName: { 'Fn::GetAtt': [Match.anyValue(), 'Name'] },
+    FunctionVersion: '2018-05-29',
+    Name: 'InvokeEchofunction123123456456LambdaDataSource',
+    RequestMappingTemplateS3Location: {
+      'Fn::Join': ['', ['s3://', { Ref: Match.anyValue() }, '/', { Ref: Match.anyValue() }, '/resolvers/InvokeEchofunction123123456456LambdaDataSource.req.vtl']],
+    },
+    ResponseMappingTemplateS3Location: {
+      'Fn::Join': ['', ['s3://', { Ref: Match.anyValue() }, '/', { Ref: Match.anyValue() }, '/resolvers/InvokeEchofunction123123456456LambdaDataSource.res.vtl']],
+    },
+  });
+  Template.fromJSON(stack).hasResourceProperties('AWS::AppSync::Resolver', {
+    ApiId: { Ref: Match.anyValue() },
+    FieldName: 'echo',
+    TypeName: 'Query',
+    Kind: 'PIPELINE',
+    PipelineConfig: {
+      Functions: [{ 'Fn::GetAtt': [Match.anyValue(), 'FunctionId'] }],
+    },
+    RequestMappingTemplate: Match.anyValue(),
+    ResponseMappingTemplateS3Location: {
+      'Fn::Join': ['', ['s3://', { Ref: Match.anyValue() }, '/', { Ref: Match.anyValue() }, '/resolvers/Query.echo.res.vtl']],
+    },
+  });
   expect(out.resolvers).toMatchSnapshot();
 });
 
@@ -289,11 +274,11 @@ test('two @function directives for the same lambda should produce a single datas
   expect(out.stacks).toBeDefined();
   const stack = out.stacks.FunctionDirectiveStack;
   expect(stack).toBeDefined();
-  cdkExpect(stack).to(countResources('AWS::IAM::Role', 1));
-  cdkExpect(stack).to(countResources('AWS::IAM::Policy', 1));
-  cdkExpect(stack).to(countResources('AWS::AppSync::DataSource', 1));
-  cdkExpect(stack).to(countResources('AWS::AppSync::FunctionConfiguration', 1));
-  cdkExpect(stack).to(countResources('AWS::AppSync::Resolver', 2));
+  Template.fromJSON(stack).resourceCountIs('AWS::IAM::Role', 1);
+  Template.fromJSON(stack).resourceCountIs('AWS::IAM::Policy', 1);
+  Template.fromJSON(stack).resourceCountIs('AWS::AppSync::DataSource', 1);
+  Template.fromJSON(stack).resourceCountIs('AWS::AppSync::FunctionConfiguration', 1);
+  Template.fromJSON(stack).resourceCountIs('AWS::AppSync::Resolver', 2);
 });
 
 test('two @function directives for the same field should be valid', () => {
@@ -312,18 +297,17 @@ test('two @function directives for the same field should be valid', () => {
   expect(out.stacks).toBeDefined();
   const stack = out.stacks.FunctionDirectiveStack;
   expect(stack).toBeDefined();
-  cdkExpect(stack).to(countResources('AWS::AppSync::Resolver', 1));
-  cdkExpect(stack).to(
-    haveResource('AWS::AppSync::Resolver', {
-      ApiId: { Ref: anything() },
+  Template.fromJSON(stack).resourceCountIs('AWS::AppSync::Resolver', 1);
+  Template.fromJSON(stack)
+    .hasResourceProperties('AWS::AppSync::Resolver', {
+      ApiId: { Ref: Match.anyValue() },
       FieldName: 'echo',
       TypeName: 'Query',
       Kind: 'PIPELINE',
       PipelineConfig: {
-        Functions: [{ 'Fn::GetAtt': [anything(), 'FunctionId'] }, { 'Fn::GetAtt': [anything(), 'FunctionId'] }],
+        Functions: [{ 'Fn::GetAtt': [Match.anyValue(), 'FunctionId'] }, { 'Fn::GetAtt': [Match.anyValue(), 'FunctionId'] }],
       },
-    }),
-  );
+    });
 });
 
 test('@function directive applied to Object should throw Error', () => {