+++ packages/amplify-graphql-index-transformer/src/resolvers.ts
@@ -1,9 +1,13 @@
 import { generateApplyDefaultsToInputTemplate } from '@aws-amplify/graphql-model-transformer';
-import { MappingTemplate, GraphQLTransform, AmplifyApiGraphQlResourceStackTemplate, SyncUtils, StackManager } from '@aws-amplify/graphql-transformer-core';
-import { DataSourceProvider, StackManagerProvider, TransformerContextProvider, TransformerPluginProvider, TransformerResolverProvider } from '@aws-amplify/graphql-transformer-interfaces';
-import { DynamoDbDataSource } from '@aws-cdk/aws-appsync';
-import { Table } from '@aws-cdk/aws-dynamodb';
-import * as cdk from '@aws-cdk/core';
+import {
+  MappingTemplate, GraphQLTransform, AmplifyApiGraphQlResourceStackTemplate, SyncUtils, StackManager,
+} from '@aws-amplify/graphql-transformer-core';
+import {
+  DataSourceProvider, StackManagerProvider, TransformerContextProvider, TransformerPluginProvider, TransformerResolverProvider,
+} from '@aws-amplify/graphql-transformer-interfaces';
+import { DynamoDbDataSource } from '@aws-cdk/aws-appsync-alpha';
+import { Table } from 'aws-cdk-lib/aws-dynamodb';
+import * as cdk from 'aws-cdk-lib';
 import { Kind, ObjectTypeDefinitionNode, TypeNode } from 'graphql';
 import {
   and,
@@ -651,8 +655,8 @@ function setSyncQueryFilterSnippet(deltaSyncTableTtl: number) {
     compoundExpression([
       set(ref('filterArgsMap'), ref('ctx.args.filter.get("and")')),
       generateDeltaTableTTLCheck(
-        'isLastSyncInDeltaTTLWindow', 
-        deltaSyncTableTtl, 
+        'isLastSyncInDeltaTTLWindow',
+        deltaSyncTableTtl,
         'ctx.args.lastSync'
       ),
       ifElse(
@@ -691,8 +695,8 @@ function setSyncQueryFilterSnippet(deltaSyncTableTtl: number) {
 }
 
 const generateDeltaTableTTLCheck = (
-  deltaTTLCheckRefName: string, 
-  deltaTTLInMinutes: number, 
+  deltaTTLCheckRefName: string,
+  deltaTTLInMinutes: number,
   lastSyncRefName: string
 ): Expression => {
   const deltaTTLInMilliSeconds = deltaTTLInMinutes * 60 * 1000;
@@ -893,9 +897,10 @@ export const getDeltaSyncTableTtl = (resourceOverrides: $TSAny, resource: Transf
 }
 
 export const getResourceOverrides = (transformers: TransformerPluginProvider[], stackManager?: StackManagerProvider | null): $TSAny => {
-  try {
+  if (stateManager.currentMetaFileExists(undefined)) {
     const meta = stateManager.getCurrentMeta(undefined, { throwIfNotExist: false });
-    const gqlApiName = _.entries(meta?.api).find(([, value]) => (value as { service: string }).service === 'AppSync')?.[0];
+    const gqlApiName = _.entries(meta?.api)
+      .find(([, value]) => (value as { service: string }).service === 'AppSync')?.[0];
     if (gqlApiName && stackManager) {
       const backendDir = pathManager.getBackendDirPath();
       const overrideDir = path.join(backendDir, 'api', gqlApiName);
@@ -909,9 +914,6 @@ export const getResourceOverrides = (transformers: TransformerPluginProvider[], 
       });
       return localGraphQLTransformObj.applyOverride(stackManager as StackManager);
     }
-  } catch (error) {
-    // do not throw but return empty overrides
-    return {};
   }
   return {};
 }