+++ packages/amplify-graphql-model-transformer/src/graphql-model-transformer.ts
@@ -35,11 +35,11 @@ import {
   StreamViewType,
   Table,
   TableEncryption,
-} from '@aws-cdk/aws-dynamodb';
-import * as iam from '@aws-cdk/aws-iam';
-import { CfnRole } from '@aws-cdk/aws-iam';
-import * as cdk from '@aws-cdk/core';
-import { CfnDataSource } from '@aws-cdk/aws-appsync';
+} from 'aws-cdk-lib/aws-dynamodb';
+import * as iam from 'aws-cdk-lib/aws-iam';
+import { CfnRole } from 'aws-cdk-lib/aws-iam';
+import * as cdk from 'aws-cdk-lib';
+import { CfnDataSource } from 'aws-cdk-lib/aws-appsync';
 import {
   DirectiveNode,
   FieldDefinitionNode,
@@ -1237,6 +1237,10 @@ export class ModelTransformer extends TransformerModelBase implements Transforme
       ...(context.isProjectUsingDataStore() ? { timeToLiveAttribute: '_ttl' } : undefined),
     });
     const cfnTable = table.node.defaultChild as CfnTable;
+
+    // CDK started to append hash to logical id of dynamodb table.
+    // This line overrides that behavior to avoid deletion and re-creation of existing tables.
+    cfnTable.overrideLogicalId(tableLogicalName);
 
     cfnTable.provisionedThroughput = cdk.Fn.conditionIf(usePayPerRequestBilling.logicalId, cdk.Fn.ref('AWS::NoValue'), {
       ReadCapacityUnits: readIops,