+++ packages/amplify-graphql-transformer-interfaces/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-interfaces@2.1.0-beta.4...@aws-amplify/graphql-transformer-interfaces@2.1.0) (2023-03-01)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-interfaces
+
+# [2.1.0-beta.6](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-interfaces@2.1.0-beta.4...@aws-amplify/graphql-transformer-interfaces@2.1.0-beta.6) (2023-02-21)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-interfaces
+
+# [2.1.0-beta.5](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-interfaces@2.1.0-beta.4...@aws-amplify/graphql-transformer-interfaces@2.1.0-beta.5) (2023-02-15)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-interfaces
+
+# [2.1.0-beta.4](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-interfaces@1.14.12...@aws-amplify/graphql-transformer-interfaces@2.1.0-beta.4) (2023-02-03)
+
+### Bug Fixes
+
+- lock CDK v2 version ([#923](https://github.com/aws-amplify/amplify-category-api/issues/923)) ([2afe40c](https://github.com/aws-amplify/amplify-category-api/commit/2afe40cf13e7d1ee7db37988b9b3297768c7bd0a))
+
+### Features
+
+- migrate index and model transformers to CDK v2 ([#860](https://github.com/aws-amplify/amplify-category-api/issues/860)) ([886ab6c](https://github.com/aws-amplify/amplify-category-api/commit/886ab6c1eb699f9a09f273c76b3c419c73004f9b))
+
+# [2.1.0-beta.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-interfaces@2.1.0-beta.0...@aws-amplify/graphql-transformer-interfaces@2.1.0-beta.3) (2022-12-27)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-interfaces
+
+# [2.1.0-beta.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-interfaces@2.1.0-beta.0...@aws-amplify/graphql-transformer-interfaces@2.1.0-beta.2) (2022-12-12)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-interfaces
+
+# [2.1.0-beta.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-interfaces@2.1.0-beta.0...@aws-amplify/graphql-transformer-interfaces@2.1.0-beta.1) (2022-11-30)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-interfaces
+
+# [2.1.0-beta.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-interfaces@2.1.0-cdkv2.2...@aws-amplify/graphql-transformer-interfaces@2.1.0-beta.0) (2022-11-18)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-interfaces
+
+# [2.1.0-cdkv2.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-interfaces@1.14.9...@aws-amplify/graphql-transformer-interfaces@2.1.0-cdkv2.2) (2022-11-15)
+
+### Bug Fixes
+
+- lock CDK v2 version ([#923](https://github.com/aws-amplify/amplify-category-api/issues/923)) ([2afe40c](https://github.com/aws-amplify/amplify-category-api/commit/2afe40cf13e7d1ee7db37988b9b3297768c7bd0a))
+
+### Features
+
+- migrate index and model transformers to CDK v2 ([#860](https://github.com/aws-amplify/amplify-category-api/issues/860)) ([886ab6c](https://github.com/aws-amplify/amplify-category-api/commit/886ab6c1eb699f9a09f273c76b3c419c73004f9b))
+
+# [2.1.0-cdkv2.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-interfaces@2.1.0-cdkv2.0...@aws-amplify/graphql-transformer-interfaces@2.1.0-cdkv2.1) (2022-10-24)
+
+### Bug Fixes
+
+- lock CDK v2 version ([#923](https://github.com/aws-amplify/amplify-category-api/issues/923)) ([2afe40c](https://github.com/aws-amplify/amplify-category-api/commit/2afe40cf13e7d1ee7db37988b9b3297768c7bd0a))
+
+# [2.1.0-cdkv2.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-interfaces@1.14.7...@aws-amplify/graphql-transformer-interfaces@2.1.0-cdkv2.0) (2022-10-24)
+
+### Features
+
+- migrate index and model transformers to CDK v2 ([#860](https://github.com/aws-amplify/amplify-category-api/issues/860)) ([886ab6c](https://github.com/aws-amplify/amplify-category-api/commit/886ab6c1eb699f9a09f273c76b3c419c73004f9b))
+
 ## [1.14.12](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-interfaces@1.14.11...@aws-amplify/graphql-transformer-interfaces@1.14.12) (2023-01-26)
 
 **Note:** Version bump only for package @aws-amplify/graphql-transformer-interfaces