+++ packages/amplify-schema-validator/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-validation@0.1.1...@aws-amplify/graphql-schema-validation@0.2.0) (2023-02-27)
+
+### Bug Fixes
+
+- fixed coverage ([cb12702](https://github.com/aws-amplify/amplify-category-api/commit/cb127021105e4a114cb862c29a22ca84491aab88))
+- fixed index should be non scalar case for list type ([6ed2359](https://github.com/aws-amplify/amplify-category-api/commit/6ed2359b146d5ce24595c46acefcc1c10b12b10a))
+- minor refactoring ([9d99b6c](https://github.com/aws-amplify/amplify-category-api/commit/9d99b6c9cb2068512aa0bac6b2c8545ad0b5c57e))
+
+### Features
+
+- added schema validators ([2c8c136](https://github.com/aws-amplify/amplify-category-api/commit/2c8c1362fa7410d8fd0e67bbe8c7af61f456c2c9))
+
 ## [0.1.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-validation@0.1.0...@aws-amplify/graphql-schema-validation@0.1.1) (2023-02-10)
 
 ### Bug Fixes